Update the bytecode after each analysis step
Summary:
Prior to this diff, the analysis step would decide what changes to make, and analyze those changes, but then throw everything away.

This changes things so that we keep a set of changes for each block, and then apply them during the update phase. This also structures things so that at any time you can see the full list of bytecodes (after optimization) that precede the current bytecode. Together with the previous diff, this will give us the ability to optimize groups of instructions in the interpreter.

This diff doesn't add any such optimizations, and stops calling the existing `group` analysis steps. Follow ups will add peephole optimizations, and a replacement for the existing groups.

This also changes the responsibility for optimizing in the presence of jmp_setdest(). If you call jmp_setdest you should also optimize the bytecode (eg for a Jmp[N]Z, reduce it to a PopC). And if you optimize away a conditional branch because it always falls through, there's no need to call jmp_nevertaken - since there's no longer a branch at the end of the block. If you leave the bytecode in place (eg IterInit), the optimizer will take care of updating any branches that *don't* go to the specified destination to point to an unreachable block.

Reviewed By: jano

Differential Revision: D14767156

fbshipit-source-id: 3c41df710bbb5a978d95210e37a0680ff2412c51