Add JitPgoPseudomain command line option to allow retranslation.
This new flag changes the behavior of shouldPGOFunc() additionally
allowing pseudomains to be PGO'd. This allows pseudomains to be
retranslated as a TransOptimize translation instead of the a simple
TransLive translation. This leads to more optimized pseudomains,
but more importantly, any call sites within the pseudomain that
originally pointed to a TransProfile translation will now point to
a TransOptimize translation.

Experiments using tcprint on the OSS Mediawiki target showed
a significant number of perf samples in aProf without the new
JitPgoPseudomain flag. Setting the existing JitPseudomain to false
led to profiles with no perf samples in aProf because the
pseudomains were always always interpretted and thus the call sites
were never "stale". The performance was mildly better, but
intpretting pseudomains was slow. Using the new JitPgoPseudomain
gave the strongest performance increase also with no perf samples
in aProf.