Optimize CombineAndResolveTypeStruct instruction in HHBBC
Summary:
This diff adds an optimizations to `CombineAndResolveTypeStruct` such that if there is only one type structure (i.e. no need to do any combining) and if this type structure either does not need any resolution or we can resolve it statically, then do that.

This diff also revealed an inconsistency in type structure resolution.

Reviewed By: paulbiss

Differential Revision: D14850255

fbshipit-source-id: a3f9c4274be7024100bf60c36103b282c3f453e7