Make the primary error message more prominent
Summary:
The contextual error formatter does not guarantee that the topmost line is the primary offending line.

Ensure that the primary error is shown in a prominent colour, making it easier to spot where the primary problem is. Use a dim colour for secondary messages.

Preserve files in the order that occurred in the original error, so the primary error is always in the first file shown.

Refactor supporting functions to clarify the relationship between the different formatting helper functions.

Reviewed By: andrewjkennedy

Differential Revision: D14831148

fbshipit-source-id: 7ffd2fa00072d4dca6b5305adc51f366baf8b15f