anonymous function should not capture variables not in 'use' clause.
Summary:
Currently, for lambdas and anonymous functions, we stash all continuations but Next. In this diff, we also stash the Next continuation for anonymous functions and reinject variables from the use clause.

This requires to add a boolean to Efun to tell whether that's an anonymous function (not fond of this solution because that adds up to the Lfun Efun mess, so any suggestion welcome).

There are also some not so interesting changes in the helper functions in Typing_env, Typing_lenv, Typing_lenv_cont.

Reviewed By: andrewjkennedy

Differential Revision: D14664890

fbshipit-source-id: d8b62b16ba169a3c285b7a14075c4be5692ec503