Remove some now-dead code
Summary:
This was for keeping track of fpi stack modifications when the
fpush/fcall was optimized away, or converted to FCallBuiltin.

It was only needed if the fpush and fcall were in different
blocks. But thats no longer possible. Also add an assertion that the
propagated fpi stack is always empty.

Reviewed By: ricklavoie

Differential Revision: D14898213

fbshipit-source-id: 3c2318eb61f4a79a811b3f0a167aa34d95e03ba2