Use half the cpu workers when possible
Summary:
When D14910086 makes it to Sandcastle we no longer restrict the cpu's on the command line.

Instead, I added a hook to Hack that let's us modify the worker count when running Zoncolan. This way, we are able to also speed up Hack's parsing (which slows down ~30s when using too many workers).

Reviewed By: zlandau

Differential Revision: D14910647

fbshipit-source-id: 1f32693ec8f68159dcfdee4b3eddac00b8f8c111