SourceText, LexableTrivia, LexableToken
Summary:
Some basic type signatures (as we would call them in OCaml) or traits (as they are in Rust).

The major controversy of this diff is representing source text as vector of bytes instead of as string. Rust ***really*** wants you to get unicode right, and it's commendable standard we should aspire to meet one day, but it's not something the current implementation does. Switching implementation languages and the meaning of the phrase “character at index i” at the same time sounds like a recipe for disaster, so I'm trying to maintain the current behavior, which is annoying with str type - hence `&[u8]`. It still forces us to do `from_utf8_unchecked` from time to time when interfacing with string literals, which have mismatched type now.

As noted by Jake, we could probably drop having to support invalid utf sequences and theoretically use `&str`, but then (since we don't want to change the indexing rules yet), we end up dropping it back to raw byte sliced over and over in lexer.

On the other hand, when we get to printing stuff (say, for parser errors), this gets inconvenient in other direction. In the end, we'll likely need to have both kinds of accessors.

 ---
The other, less immediately obvious controversy is `SourceText` not owning the string, but just being a wrapper around a reference to it. This is because I expect source text strings to be often existing in OCaml heap, or HHVM memory, where they cannot be managed by Rust - so `SourceText` would have to make a copy, and as we know we have some enormous files. The drawback is that it introduces lifetime management of this reference, which in turn prevents us from doing some things that in GC world just work.

For example, OCaml positioned tokens store a pointer to source text, which allows them to have a [`text`](diffusion/FBS/browse/master/fbcode/hphp/hack/src/parser/full_fidelity_positioned_token.ml;e931d3193229e49e78a4d82b7a5bad2e231a2bc2$340) method, returning the text that token span covers. I tried tying the token lifetime to source text lifetime, but that just exploded. Instead, we can make the `text` method accept source text, and let caller worry about having both available. In practice, in all three places that we encountered so far this was quiet easy to achieve.

(Note: this ignores all push blocking failures!)

Reviewed By: losvald

Differential Revision: D14195683

fbshipit-source-id: 39eed7e8119438e7e3d47692ebd1a0dc8736f57b