(basic) Smart constructors
Summary:
This is the simplest possible version of smart constructors. It's just enough to be able to construct syntax tree, but it's missing a lot of other stuff (threading of state, not forcing users to implement NodeType trait, design that makes them "inheritable"). Leo has those things implemented in dependent diffs, so don't pay too much attention to design in this diff - it's only a temporary scaffolding.

(Note: this ignores all push blocking failures!)

Reviewed By: losvald

Differential Revision: D14195772

fbshipit-source-id: 0e1e1ab6b089e94a40d3f66da0e1e25200525bc0