Lexer
Summary:
Line by line from `full_fidelity_lexer.ml` The interesting part is lexers being mutable and backtracking done by `clone`-ing them. This is how Jake did it, and the motivation is reducing the number of allocations (which, at least for this use case, are hypothesized to be cheaper in OCaml due to minor GC heap being faster than Rust allocations always having to go through (je-)malloc).

The drawback is that it makes it more possible to introduce superlinear runtime complexity - if something causes cost of cloning of the lexer to grow linearly, the whole algorithm can become quadratic. The only thing that looks like it could cause it is the vector of errors. OCaml avoids is by error list being a list, and "cloning" lists being non-op due to memory sharing. Achieving the same thing in Rust would require switching Vec to an actual reference counted linked list, which in turn seems like a lot of effort just to handle an extreme edge case (famous last words).

(Note: this ignores all push blocking failures!)

Reviewed By: losvald

Differential Revision: D14195871

fbshipit-source-id: 4e64495db707b8d01d05783ec241fb91b951f9a8