MinimalTrivia + MinimalToken + MinimalSyntax
Summary:
An observant reader can notice that all the code so far was an exercise in abstraction, with no concrete instantiation of any of the types described. Time to change it.

`minimal_trivia.rs` is the best example of my beef with `rustfmt` - this could be so much more concise!

(Note: this ignores all push blocking failures!)

Reviewed By: losvald

Differential Revision: D14195931

fbshipit-source-id: 5dbcfce34e70fe193ac806689fcec14f0af0e354