Rewrite logic for inheritance of typeconsts in terms of precomputed abstract
Summary:
Instead of recomputing based on the presence or absence of types or constraints and deciding whether to inherit, this spells out the inheritance rules based on the abstract_kind during the Aast generation.

This is more easily extensible to abstract type constants with defaults, and I feel it is clearer than the boolean logic.

Reviewed By: Matt-Schellhas

Differential Revision: D14794324

fbshipit-source-id: 09817c67184bb6dd7c9aee9d1cfe20df53dda7c8