Treat array as subtype of Traversable<any>
Summary:
Currently, the untyped array type `array` is a subtype of `Traversable<t>` and `KeyedTraversable<t,u>` for *any* `t` and `u`.

This diff makes a small change: it is now treated as a subtype of `Traversable<_>` and `KeyedTraversable<_,_>` where `_` is the "any" type. This leads to better propagation of `_`, rather than instantiation with `[unresolved]` (under legacy checking) or `nothing` (in new inference), or unresolved type variables that later produce errors (new inference). See the new test for an example of this latter behaviour.

Reviewed By: dlreeves

Differential Revision: D14916991

fbshipit-source-id: 66d044a0bcf5462b60c3de5ba2b1bc2f056c9862