replace Local_id.get with Local_id.make_unscoped
Summary:
Somehow when fixing the control flow in naming I missed these calls to Local_id.get for function parameters, which contrary to locals are not annotated with a local id. Just like other locals, function parameters do not need to be assigned any local id and can just receive id 0, since the proper handling of control flow in Typing will figure out their type correctly. Local_id.get however returns an id based on the hash of the string name of the variable. This used to be perhaps useful, but is now completely arbitrary.

The previous buggy behaviour due to Local_id.get returning non-zero was only visible in lambdas, in cases like the added test cases.

Reviewed By: kmeht

Differential Revision: D14682827

fbshipit-source-id: bedc30db6ce99ba949dcff89a14605e2a6c0c8c4