New inference: better treatment of union in coalesce operator, switch
Summary:
Changed our checking of the `??` coalesce operator to simply mimic what we would do with a call to an equivalent method. This means that our generic constraint solving and union simplification code kicks in.

But this still doesn't simplify under nullables. This affects `switch`, whose check for exhaustiveness is sensitive to unions. So simplify unions in the type of the switch scrutinee.

Reviewed By: CatherineGasnier

Differential Revision: D14891316

fbshipit-source-id: 9fa34c237ed37f71e3615d9808fa1dd1ac7c1849