Use the default busy handler
Summary:
I experimented with various busy handler schemes. It was possible to
make it a little better, but it was always easy to trigger the "failed
to get lock" warning. The default busy handler's back off scheme seems
to be about as good as any, and the warning isn't really helpful - if
we're trying to write a unit, failure is ignored (we'll try again next
time we load the unit), and for other things, we'll report an error
anyway.

Reviewed By: ricklavoie, binliu19

Differential Revision: D14930670

fbshipit-source-id: b6a4efc2b17986492a799d1a77bf35b25c087f7b