Rename Typing_class_member_cache to Lazy_string_table
Summary:
I gave Typing_class_member_cache that name because that was the purpose for which I wrote it, and I was not feeling very creative. It is not used only for class members, though--the very first diff making use of it uses it to store class ancestors ({D13571853}). Since it needs to be used by both decl and typing, I found that I needed to put it in the typing_defs Buck target/Dune library, which also doesn't seem right to me.

Instead, I'm suggesting that we call this data structure Lazy_string_table, and store it in src/utils/collections (solving the dependency issue which led me to initially put it in the typing_defs target).

Reviewed By: ljw1004

Differential Revision: D14889701

fbshipit-source-id: f0cb20defaf61ffa98031b71b55d13f336c7770a