Delete Repo copy constructor
Summary:
My programs kept segfaulting while I worked on an addition to HHVM's `.hhbc` repo. After four days of commenting and uncommenting out code, I narrowed the problem down to one line of code:

```
{
  auto repo = Repo::get();
}
```

Which copy-constructs the repo and then runs the destructor, destroying the SQLite connection. Instead, I should have written

```
{
  auto& repo = Repo::get();
}
```

But the real fix is to follow the rule of five here so that the compiler will catch problems like this for future me.

Reviewed By: alexeyt, markw65

Differential Revision: D14856118

fbshipit-source-id: 0cc91e8db46f3529f012da46171d520e27e57d3d