move some preprocessor macros to a separate file
Summary: A few macros, such as `USE_JEMALLOC`, `USE_JEMALLOC_EXTENT_HOOKS`, along with some global constants, are moved to a separate file, alloc-defs.h, so that it can be included in a few allocator-related files.  With this, I no longer have to place declarations in unnatural places to handle cyclic inclusion among headers.  It also makes it easier to not break mac build, etc.

Reviewed By: markw65

Differential Revision: D14929469

fbshipit-source-id: 9e1c38693ebeb35c269eeb7c70d6bb99b3972f2f