Support abstract const type default syntax
Summary:
Tiny change that parses the abstract-with-default syntax as a standard partially abstract type constant. All the rest of the plubming is in place throughout the typechecker.
I rewrote the Ast_to_nast bit to trace the kind of the type constant while retaining all of the existing functionality.

For now, the runtime implementation as a partially abstract constant is fine, but long term they will begin to error on access.

Reviewed By: kmeht

Differential Revision: D14788277

fbshipit-source-id: fefb32bea105cef7649cc9682d4e4359f129b5e4