Pull statement into Awaitall
Summary:
Today we emit the Awaitall separate from the statement the temp vars get injected into. This has the downside that we can't nicely scope the temp vars to just the next statement.

This diff pulls the statement into the Awaitall and removes the need for another set of tmps as well as scoping them to just these two statements.

I'm unsure how to properly use `Scope.with_unnamed_locals`. Jan any help on that would be great :)

Reviewed By: paulbiss, jano

Differential Revision: D14823764

fbshipit-source-id: 6ec4451cecf7bd96256bb0cc0d6f5fa381675e32