Improve expected errors for tests.
Summary:
A lot of test expected files were simply testing for "Fatal error:" following by anything - which means that a broken change that still emits "Fatal error:" would pass the test.  Tests should be more specific about the error wording expected.

Most of these changes were mechanically produced using the output of hhvm and post-processing to remove specific directory paths.

The intention is to tighten the expected messages now and if we decide to loosen them later we can do so on a case-by-case basis - although we should really be at least mentioning *some* pertinent words that the error should expect.

Differential Revision: D14929831

fbshipit-source-id: 68577ca63d7aa99c6325c7cef0d910173e85b3b5