New inference: do not apply key type checks for array update
Summary: A previous diff D14520917 introduced TAST checks for bad key types in array indexing. Unfortunately it applied these checks also to array updates, which is too restrictive: given that key types are covariant, we want to allow "larger" index types on array update. This diff fixes the check so that it applies only to rvalue checks, not those in lvalue (assignment) positions.

Reviewed By: kmeht

Differential Revision: D14703728

fbshipit-source-id: 7ed2b7be73b8441aff766cf183933f39bde9ad07