Back out "Avoid examining folded methods table when declaring condition_types"
Summary:
Reverts {D13568133}.

When I shipped that diff, I did not fully understand the potential cost of fetching the shallow declaration of every ancestor. I supposed that, since we provided a worker-local cache of shallow declarations, and declared every ancestor before declaring a class, we could generally expect shallow declarations of ancestors to be present in the cache. The perf results in the test plan indicate that this frequently not the case--ancestors are often declared on other workers, and attempting to fetch their shallow declaration when it is not present in the cache involves an expensive re-parse.

With the gating changes I have made, D13568133 is no longer necessary to unblock shallow class declaration, so this diff reverts it.

Reviewed By: arxanas

Differential Revision: D14938829

fbshipit-source-id: dba7cb04a9c249b9adf3ffe43d8ec7f6da635e49