Basic typechecker support for like-types
Summary:
This diff adds a bare-bones implementation of like-types to the typechecker, in which they're defined simply as the union of `dynamic` and the given type. This doesn't give us all the behavior we want yet (see TODOs in tests), but does bring us one step closer.

Following diffs will:
- Improve how like-types are printed (some of the error messages are fairly confusing).
- Actually implement the coercion behavior required to make like-types act as we want at enforcement points
- Implement explicit coercion via the `as` operator

Reviewed By: vassilmladenov

Differential Revision: D14942402

fbshipit-source-id: 2d90bcb274d42051fe9695314d688f70f751ddfa