Stop binding refs into arrays in SOAP
Summary: As with mysqli there was no reason that we needed to store these refs into an `Array`. In fact in this case we were using one array to map from `xmlNodePtr` to `RefData`/`ObjectData` **and** using the same array as a reverse mapping (casting pointers to `int64_t` in the process). This diff switches to `req` containers and keys on `HeapObject*` instead so that these roots are visible to the GC.

Reviewed By: ricklavoie

Differential Revision: D14884579

fbshipit-source-id: e0fab70efdc783a8c42de3004d6a7637cf4a6e7a