Unified symbol search API that works with multiple providers
Summary:
Begin scaffolding to enable the symbol search system to implement multiple providers with a single consistent API (index_search).

This new API does not depend on any of the trie-specific features, and it can separate "local file symbols" from "global file symbols".

In this current diff nothing is yet switched to use it; that will follow in the next diff.

Add a test suite that can be used to test different autocomplete providers against each other.

Reviewed By: 2BitSalute

Differential Revision: D14900393

fbshipit-source-id: 4880d5f02f08284fa0372242b2776ff67631aba9