Add shared memory cache for linearizations
Summary: While experimenting with running D13571860 over WWW, I found that sharing linearizations between workers was critical for performance--this diff reduced a 74% perf regression (from trunk) to a 24% regression. This diff adds a shared memory cache for linearizations so that completed linearizations (we cannot serialize the functional values in partially-computed linearizations) can be shared between workers.

Reviewed By: pittsw

Differential Revision: D13571859

fbshipit-source-id: 98782b23a74f60444b3fbff59537ae47035293d2