Move enforceable override handling to Typing_extends
Summary: If we propagate ttc_typeconst to overrides, we end up with a severe performance regression with shallow class declarations (on the order of 15%). Instead of propagating this property to overrides and doing this enforcement in a Tast_check, do the enforcement in Typing_extends, so that propagating the property to overrides is not necessary.

Reviewed By: arxanas

Differential Revision: D14567951

fbshipit-source-id: d64425328af21272eb0f851cabb81aa97f154705