Use tv_val in collections interfaces
Summary:
This diff is a prerequisite for making layout changes to existing collections. Outside of wide mode, it shouldn't have any performance affects.

Most of the changes here update collections.h to return tv_lvals and propagate them through the rest of the codebase.

I also modified a few of the collections-vector methods to use tv_lval internally with a new dataAt(int64_t) method which we can easily switch to a new layout. The bulk operations (map, filter, popFront, etc.) still use the data() method and make layout assumptions.

Reviewed By: mofarrell

Differential Revision: D14748846

fbshipit-source-id: fc6bb8563073ec19a20cd766fade7c11c46fe1ef