Make Vector collection layout-agnostic
Summary:
This diff makes almost all the code in ext_collections-vector.cpp agnostic to the layout of its backing store, PackedArray.

Most of the changes are in various bulk operations (map, filter, take, etc.) that iterate over the array. Many of them now use a copySlice helper that we can optimize with layout-specific logic if needed.

There are two places where I left the layout assumption in place. Both are marked with static asserts:
- resizeImpl (uses memcopy to bulk-copy data from the old buffer to the new one)
- removeKeyImp (uses memmove to bulk-move data; I made popFirst share this code)

If we use a new layout, we can probably still do the memcopy on resizeImpl, just with a different size in bytes. For example, if we use the "8 type bytes + 8 value words" layout, an n-element vector will use `n + ((n + 7) >> 3)` bytes. Using memmove for removeKeyImpl is a lot trickier with that layout, though there are still ways to do it.

Outside of the core changes, I added a few utilities like tvSwap and new Variant constructors. I think tvSwap is useful elsewhere, and I may try to make that change in another diff. The Variant constructors are a little rough, though - for instance, I wasn't able to call them with a TypedValue, and I couldn't get them to work with T&& instead.

Reviewed By: mofarrell

Differential Revision: D14762565

fbshipit-source-id: e288b3eece5d6733d4a30b8391c03a4720254e7e