Let each unit decide whether to use the litstr table
Summary:
I want to be able to use a repo to cache units for `--hphp`
builds. The problem at present is that when Option::WholeProgram is
true, units that get created always use the global litstr table; and
when RuntimeOption::RepoAuthoritative is true, units loaded from the
repo assume they're getting global ids.

We can't (easily) build a cache incrementally if it uses the global
litstr table, but we also can't feed a mix of units to hhbbc where
some use the litstr table, and some do not (and any newly parsed units
will use the global litstr table).

This adds a flag to each unit to say whether it was built (or should
be built) using the litstr table, and changes things so that only
units written by hhbbc use the global litstr table.

Reviewed By: ricklavoie

Differential Revision: D14919599

fbshipit-source-id: 9c33b99f43f0f86c202bebb1b4aeb62fbbefc6ad