Move symbol-refs to UnitEmitter
Summary:
Currently, the assembler just passes symbol refs to a set of
(optional) callbacks. I want to be able to cache the result of parsing
the inputs to an hphpc build, but that means caching the symbol refs
too. The current callback mechanism doesn't even tell me which unit
the symbol refs come from, so a better solution seems to be to cache
them on the unit itself.

We don't need these for regular builds, so make it optional whether
they're generated, and in this diff, I don't provide the option to
save them to the repo. I'll add that in the diff that enables caching.

Reviewed By: paulbiss

Differential Revision: D14976841

fbshipit-source-id: 0b5e0d8d7a66f820563cbf826cb6a77943d7f547