Let repo mode builds create/use a unit cache
Summary:
If you specify a read-only local repo, `hhvm --hphp` will try to read
units from that repo, and skip parsing it if its there.

If you specify a read-write local repo, `hhvm --hphp` will
additionally write any newly parsed units back to the local repo. This
is a relatively slow process, so to avoid slowing down the build when
the cache is relatively empty, we normally just stop writing units
when parsing finishes.

Finally, if you specify --target=cache, the behavior is the same,
except hhvm will allow all the units to be written to the cache, and
won't run hhbbc, or write an optimized repo.

Reviewed By: ricklavoie

Differential Revision: D14987918

fbshipit-source-id: 2be9c4be1c57aea3ca612275d1d5cbc95a864798