Add `use ... as` and `insteadof` to the AAST
Summary:
Trait conflict resolution (https://www.php.net/manual/en/language.oop5.traits.php#language.oop5.traits.conflict) is a PHP feature that is not a Hack feature. However, despite the feature not being a Hack feature, the compiler still supports bytecode generation and the runtime still supports it. To convert codegen to use the AAST over the legacy AST and maintain parity, it is necessary to maintain support for these features in the AAST.

Note: Since neither of the new AAST nodes are necessary for typechecking, the lists are replaced with empty lists in naming and typing.

Reviewed By: arun11111

Differential Revision: D14986696

fbshipit-source-id: ee5730b05544c2dd9f971d5d562a2ee1007fa76e