PositionedSyntax
Summary:
Minimal trees are nice and easy, but in practice nobody (except `HackAST` and `LazyTrivia`) consumes them - everyone needs a more convenient Positioned tree.

This is the Rust implementation of it, which also showcases another place where differences between Rust and OCaml are apparent - while sharing of values between different structs "just works", in Rust we need to be more aware of it and strategically sprinkle `Rc` in necessary places.

(Note: this ignores all push blocking failures!)

Reviewed By: Wilfred

Differential Revision: D14630078

fbshipit-source-id: 5de25270825ff506ce6f01ce44816e91baf3e9da