better string handling in SimpleJsonParser
Summary:
HHVM has two JSON parsers, SimpleJsonParser and a state-machine-based
parser which I nicknamed the bulky parser. In most cases, it tries the simple parser
first, and, if the simple parser rejects, falls back to the bulky parser.

The simple parser wasn't handling backslashes in string literals, and some profiling
revealed that most of the time we spend in the bulky parser was because the input
had a backslash. This is sad because backslashes aren't that hard to handle.

Add support for handling backslashes (when they're encoding ASCII characters,
which is the common case.)

Perflabs show this results in a roughly 0.9% reduction in instructions, as well
as a reduction in loads and stores. After bolt, there's a roughly 1% reduction
in instructions and between 0.5% and 0.9% reduction in cpu-time-us.

Reviewed By: markw65

Differential Revision: D14819100

fbshipit-source-id: b12a026584facdd75baa0c266fadf0a795fef97d