Add state to SmartConstructors, update it in parsers
Summary:
Add state to SmartConstructors as type parameter `T` (as in OCaml), and
implement wrappers to update it without exposing this at call sites
(i.e., in declaration/expression/type/statement parser).  The former is
mostly regex search+replace, except adding state setter/getter.  The latter
is achieved via a simple macro.  (In OCaml, there are 2 SC wrappers,
which add the complexity of 2 code generators as well as 350+ methods
that are each called only a few times during parsing.)
The core logic is unit-tested in Rust [playground](https://play.rust-lang.org/?version=stable&mode=debug&edition=2018&gist=506661e8d04fd98185c27f37eac26c65).

In OCaml, there is a wrapper of SmartConstructors (SC) with State,
and another of the same signature in `flatten_smart_constructors`
that is included in `facts_smart_constructors`.  Their signatures
adds an extra state parameter in ~180 `make_` methods,
but delegates up to 12 remaining arguments to the stateless SC.
This ends up tons of generated code (and increased build time!) like:

    let make_foo arg1 arg2 ... state =
       SC.make_foo arg1 arg2 ...

This is unnecessary in Rust because a unit-like struct has no overhead
(e.g., used as map value for implementation of sets); so,
instead of creating Smart Constructor wrapper that delegates
methods of different arity (up to 12 arguments), do the opposite:
require passing a state parameter and have a dummy `NoState{}`.

Notably, this will give better optimization opportunities
for SC that *do* use a state (or equivalent), since the forwarding
can be avoided in cases where state update is a no-op
(especially if we also box all the arguments together in future).

Reviewed By: dabek

Differential Revision: D14643261

fbshipit-source-id: 4c14b29150a7dba0131f87b50484269d73a50288