Cleaner interface to LockFreePtrWrapper
Summary:
If the underlying type is a pointer, its safe to expose an interface
more like a regular smart pointer - so lets just do that.

Even for non-pointer types (such as CompactVector, or copy_ptr), we
can implement operator->() making it possible to do const operations
on the underlying object directly.

Reviewed By: binliu19

Differential Revision: D14992953

fbshipit-source-id: efd508d238a02845d063405080713c13aeb6500b