HHVM Debugger: Allow configuration to specify group permissions for debugger domain socket
Summary: This diff adds the ability to add a configuration parameter that specifies a group owner for the domain socket. If specified, the domain socket will be `chown`'ed to the specified group, and permissions will be set to -rw-rw---- so it's read/write-able only to owner and group instead of everyone.

Reviewed By: velocityboy

Differential Revision: D14096296

fbshipit-source-id: 490b6b3949bbf93758260843dd5c4720a93ba95a