Make sketchy null check a lint rather than a Hack error
Summary:
For the time being, we don't want this to be a hack error. The number of these in the codebase currently are in the 10s of thousands, they don't block hack soundness, people often do this intentionally, and there's no plan to actively work on weak typing in the near future.

Instead, just make it a lint warning.

Reviewed By: DavidSnider

Differential Revision: D15002789

fbshipit-source-id: 804095663081c28543986e1b6ec45d1870cbfbb2