Remove Box opcode
Summary:
The only place where Box opcode was emitted is when an XHP attribute is being
passed by reference (e.g. `foo(&$bar->:baz)`). Since the source is not a
reference and boxing of by-ref arguments is optional, we can just pass unboxed
value like we already do for `foo(&$bar->getAttribute('baz')` and remove the
Box opcode.

Reviewed By: paulbiss

Differential Revision: D15014179

fbshipit-source-id: b499bd59ed144fd84ba7f3d0422adfae31c3fdab