Optimize CombineAndResolveTypeStruct for nullable and soft reified generic creation
Summary:
Another one of the common TS combine and resolve cases is making a reified generic nullable or soft.
This diff adds an HHBBC optimization to this path by appending the `is_nullable` or `is_soft` to the input TS. The reified generic is always fully resolved, so all we need to do is mark the TS appropriately.

Reviewed By: markw65

Differential Revision: D14993547

fbshipit-source-id: 9bd9ec8f21cf2d3099520f68ff1c491e2da6d78e