no explicit numa binding for vm stack
Summary:
In worker mode, the thread is bound to an arena whose users are all bound to a local numa node, so the result of posix_memalign() is already in the desired node.
Furthermore, the typical size of the allocation is bigger than the jemalloc tcache threshold, so it must come from the arena.

Reviewed By: markw65

Differential Revision: D15026567

fbshipit-source-id: d7c6ff948f8905a253d72efaf40119cfc04c452c