Syntax error on duplicate child and category declarations in XHP classes
Summary:
Ensure that we treat duplicate `children ...;` and `category ...;` declarations as syntax errors.

Duplicate children declarations were previously accepted, and we would use the first one found.

Duplicate category declarations were a syntax error, but we had no tests for this.

In both cases, report the last duplicate declaration as the location of the syntax error.

One buck build is timing out.

Reviewed By: kmeht

Differential Revision: D14973620

fbshipit-source-id: 419d29ad4b7a82715814c89d21c127fe746c087c